backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
docopt==0.6.2
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
gTTS==2.5.1
gunicorn==22.0.0
idna==3.7
Js2Py==0.74
multidict==6.0.5
openai==0.28.0
packaging==24.0
pipwin==0.5.2
playsound==1.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9

pycparser==2.22
pydub==0.25.1
pygame==2.5.2
pyjsparser==2.7.1

PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
SpeechRecognition==3.10.4
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
